Update Type Preserves any Instance Attributes
Attributes on FB instances where only preserved for muxers when changing
the struct. With this change the basic configuration of the updated
instances is done centrally in the base class
AbstractUpdateFBNElementCommand. This reduces the code and simplifies
maintenance.

Furthermore the not used and wrong command ChangeSubaAppPinCommand was
removed.
